requirements.txt
gradio
openai
python-dotenv
pydub
soundfile
librosa
geocoder
pydantic
resampy
jupyter-client
ipython
ipykernel
sqlalchemy
psycopg2-binary
psycopg
sqlalchemy-utils
jinja2>=3.0.0
pgvector
openai